update_nonEn_transcripts.py now works
assumes english transcript in draft/en is the latest version, and
updates all trascript files of chosen language, or creates the
transcript files if it doesnt exist for the language.
The deepest value of the json tree (so the actual string
that will be displayed) is added as "" when its added by the script.